+++ R/global_variables.R
@@ -53,7 +53,9 @@ if (getRversion() >= "2.15.1") {
       "NAPL", "DISTANCE_RELATIVE", "MARKERS_COMMON", "id.pairwise", "input.prep",
       "CHROM_LOCUS", "NEW_MARKERS", "POP_ID_N", "INDIVIDUALS_N", "GT_N", "GT_IMP",
       "INDIVIDUALS_REP", "KEEP", "GT_VCF_NUC", "INTEGERS", "GOF", "NR", "NV",
-      "A1_NUC", "A2_NUC", "RANK", "MAF_P", "MAF_L", "CHROM", "POS", "n"
+      "A1_NUC", "A2_NUC", "RANK", "MAF_P", "MAF_L", "CHROM", "POS", "n",
+      "POP_ID_PERM", "MISSING_ORIGINAL", "MARKERS_PERM", "INDIVIDUALS_PERM",
+      "MISSING_MIX", "MAR", "MCAR", "value", "GROUPS", "IMP", "OBS", "SQUARE_ERROR"
     )
   )
 }