+++ NEWS.md
@@ -1,3 +1,8 @@
+# grur v.0.0.2 2017-04-23
+
+* work on `generate_missing` function. Still not as I wish.
+
+
 # grur v.0.0.1 2017-04-07
 
 New package is out!